web_framework/mydjango/requirements.txt
diff-match-patch==20121119
Django==1.9
django-admin-bootstrapped==2.5.7
django-dbbackup==2.3.2
django-filter==0.10.0
django-import-export==0.4.1
django-model-utils==2.4
djangorestframework==3.3.2
ecdsa==0.13
Markdown==2.6.2
MySQL-python==1.2.5
paramiko==1.16.0
pycrypto==2.6.1
Pygments==2.0.2
pytz==2015.7
requests==2.9.0
six==1.10.0
tablib==0.10.0
wechat==0.4.16
wheel==0.24.0
